.maintain/cog/FHEModels/client/requirements.txt
numpy
joblib
concrete-ml==1.6.0
fastapi
uvicorn

